Scale up the virtual trackball so it covers entire screen
Pan gestures sometimes behaved strangely because the user could
get the touch point outside the bounds of the virtual trackball.
When the touch point falls outside the track ball it gets mapped
to the circle which is the intersection between the viewing plane
and the virtual trackball sphere.  That fundamentally changes the
rotation behavior in response to dragging.  To make the behavior
consistent we need to blow up the virtual trackball so that it
covers the whole screen.  Now the user can't accidentally touch
outside of it.