Normalize the normals right away to avoid using scaled normals (Fixes #4393)
In both of our Lighting shaders we were transforming every normal with
the model transform (which is correct) but then we were failing to
normalize the result before using it in a dot product for the specular
component.  So the scale of the model transform was influencing the
specular highlights.  We were normalizing the result before using it
in a dot product for the diffuse component.  Now we normalize the
transformed normal up front so we won't accidentally use a scaled
version.